#513 Move files of the creator node
Move the files to a package de.bund.bfr.knime.fsklab.v1_9.creator to not
conflict with the ICPMF deprecated node.
